# api-docs/my-openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "植物种植助手小程序后端API文档",
    "description": "植物种植助手小程序后端服务接口文档，基于Spring Boot 2.7.18实现。",
    "contact": {
      "name": "zszleon",
      "email": "2079973876@qq.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "v1.0"
  },
  "servers": [
    {
      "url": "http://localhost:18080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "TokenAuth": []
    }
  ],
...
  "paths": {
...
    "/api/records/update": {
      "post": {
        "tags": [
          "生长记录管理"
        ],
        "summary": "更新记录",
        "description": "更新生长记录信息，需要Token认证",
        "operationId": "updateRecord",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseRecord"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "413": {
            "description": "Payload Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "TokenAuth": []
          }
        ]
      }
    },
    "/api/records/page": {
      "post": {
        "tags": [
          "生长记录管理"
        ],
        "summary": "获取记录列表",
        "description": "分页获取当前用户的生长记录列表，需要Token认证",
        "operationId": "getRecordsPage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordPageQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseRecord"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "413": {
            "description": "Payload Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "TokenAuth": []
          }
        ]
      }
    },
    "/api/records/getRecordById": {
      "post": {
        "tags": [
          "生长记录管理"
        ],
        "summary": "获取记录详情",
        "description": "根据记录ID获取生长记录详细信息，需要Token认证",
        "operationId": "getRecordById",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Id"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseRecord"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "413": {
            "description": "Payload Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "TokenAuth": []
          }
        ]
      }
    },
    "/api/records/delete": {
      "post": {
        "tags": [
          "生长记录管理"
        ],
        "summary": "删除记录",
        "description": "删除指定生长记录，需要Token认证",
        "operationId": "deleteRecord",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Id"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "413": {
            "description": "Payload Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "TokenAuth": []
          }
        ]
      }
    },
    "/api/records/createRecord": {
      "post": {
        "tags": [
          "生长记录管理"
        ],
        "summary": "创建记录",
        "description": "创建新的生长记录，需要Token认证",
        "operationId": "createRecord",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseRecord"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "405": {
            "description": "Method Not Allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "413": {
            "description": "Payload Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseObject"
                }
              }
            }
          }
        },
        "security": [
          {
            "TokenAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResponseObject": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          }
        },
        "description": "统一API响应"
      },
...
      "Id": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "业务ID",
            "format": "int64",
            "example": 123456789
          }
        },
        "description": "ID请求参数"
      },
      "ApiResponseString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "type": "string",
            "description": "响应数据"
          }
        },
        "description": "统一API响应"
      },
      "RecordRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "记录ID，更新时必填",
            "format": "int64",
            "example": 123456789
          },
          "plantId": {
            "type": "integer",
            "description": "关联植物ID",
            "format": "int64",
            "example": 123456789
          },
          "type": {
            "type": "string",
            "description": "记录类型",
            "example": "watering",
            "enum": [
              "watering",
              "fertilizing",
              "growth",
              "photo"
            ]
          },
          "recordTime": {
            "type": "string",
            "description": "记录时间",
            "format": "date-time"
          },
          "notes": {
            "type": "string",
            "description": "记录备注",
            "example": "今天浇水了"
          },
          "imageUrl": {
            "type": "string",
            "description": "图片URL（拍照记录时使用）",
            "example": "https://example.com/image.jpg"
          }
        },
        "description": "生长记录请求DTO"
      },
      "ApiResponseRecord": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "$ref": "#/components/schemas/Record"
          }
        },
        "description": "统一API响应"
      },
      "Record": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "plantId": {
            "type": "integer",
            "format": "int64"
          },
          "type": {
            "type": "string"
          },
          "recordTime": {
            "type": "string",
            "format": "date-time"
          },
          "notes": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          }
        },
        "description": "响应数据"
      },
      "RecordPageQueryRequest": {
        "type": "object",
        "properties": {
          "pageNum": {
            "minimum": 1,
            "type": "integer",
            "description": "页码",
            "format": "int32",
            "example": 1
          },
          "pageSize": {
            "minimum": 1,
            "type": "integer",
            "description": "每页条数",
            "format": "int32",
            "example": 10
          },
          "plantId": {
            "type": "integer",
            "description": "植物ID（可选）",
            "format": "int64",
            "example": 123456789
          }
        },
        "description": "生长记录分页查询请求"
      },
      "ApiResponsePageResponseRecord": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "响应码",
            "format": "int32",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "操作成功"
          },
          "data": {
            "$ref": "#/components/schemas/PageResponseRecord"
          }
        },
        "description": "统一API响应"
      },
      "PageResponseRecord": {
        "type": "object",
        "properties": {
          "list": {
            "type": "array",
            "description": "数据列表",
            "items": {
              "$ref": "#/components/schemas/Record"
            }
          },
          "total": {
            "type": "integer",
            "description": "总记录数",
            "format": "int64",
            "example": 100
          },
          "pageNum": {
            "type": "integer",
            "description": "当前页码",
            "format": "int32",
            "example": 1
          },
          "pageSize": {
            "type": "integer",
            "description": "每页条数",
            "format": "int32",
            "example": 10
          },
          "totalPages": {
            "type": "integer",
            "description": "总页数",
            "format": "int32",
            "example": 10
          }
        },
        "description": "分页响应"
      },
...
    },
    "securitySchemes": {
      "TokenAuth": {
        "type": "apiKey",
        "description": "用户认证Token，通过微信登录接口获取",
        "name": "X-Token",
        "in": "header"
      }
    }
  }
}